@@ -1,70 +1,72 @@
 {
-  "name": "starter-restaurant-reservation-front-end",
-  "version": "1.0.0",
-  "description": "Starter code for restaurant reservation frontend",
-  "main": "src/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Thinkful-Ed/starter-restaurant-reservation",
-    "directory": "/front-end"
-  },
-  "private": true,
-  "engines": {
-    "node": "16"
-  },
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "web-vitals": "^0.2.4"
-  },
-  "scripts": {
-    "build": "react-scripts build",
-    "eject": "react-scripts eject",
-    "start": "react-scripts --openssl-legacy-provider start",
-    "test": "npx cross-env CI=true npm run test:run",
-    "test:run": "run-s test:unit test:e2e",
-    "test:1": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-01\"",
-    "test:2": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-02\"",
-    "test:3": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-03\"",
-    "test:4": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-04\"",
-    "test:5": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-05\"",
-    "test:6": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-06\"",
-    "test:7": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-07\"",
-    "test:8": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-08\"",
-    "test:e2e": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* test:e2e:run",
-    "test:e2e:run": "export NODE_OPTIONS=--openssl-legacy-provider && wait-on http://localhost:5001/reservations?date=2020-01-01 && jest --detectOpenHandles --forceExit --config ./e2e/jest.config.js",
-    "test:start:backend": "npm run test:start --prefix ./../back-end",
-    "test:start:frontend": "npx cross-env BROWSER=none NODE_ENV=test react-scripts --openssl-legacy-provider start",
-    "test:unit": "react-scripts test"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "cross-fetch": "^3.0.6",
-    "npm-run-all": "^4.1.5",
-    "jest-puppeteer": "^6.0.0",
-    "puppeteer": "^10.4.0"
-  }
+    "name": "starter-restaurant-reservation-front-end",
+    "version": "1.0.0",
+    "description": "Starter code for restaurant reservation frontend",
+    "main": "src/index.js",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/Thinkful-Ed/starter-restaurant-reservation",
+        "directory": "/front-end"
+    },
+    "private": true,
+    "engines": {
+        "node": "16"
+    },
+    "dependencies": {
+        "@tanstack/react-query": "^5.17.19",
+        "@testing-library/jest-dom": "^5.11.9",
+        "@testing-library/react": "^11.2.3",
+        "@testing-library/user-event": "^12.6.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-router": "^5.2.0",
+        "react-router-dom": "^5.2.0",
+        "react-scripts": "5.0.1",
+        "web-vitals": "^0.2.4"
+    },
+    "scripts": {
+        "build": "react-scripts build",
+        "eject": "react-scripts eject",
+        "start": "react-scripts start",
+        "test": "npx cross-env CI=true npm run test:run",
+        "test:run": "run-s test:unit test:e2e",
+        "test:1": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-01\"",
+        "test:2": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-02\"",
+        "test:3": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-03\"",
+        "test:4": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-04\"",
+        "test:5": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-05\"",
+        "test:6": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-06\"",
+        "test:7": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-07\"",
+        "test:8": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* \"test:e2e:run -- us-08\"",
+        "test:e2e": "export NODE_OPTIONS=--openssl-legacy-provider && run-p -l -r test:start:* test:e2e:run",
+        "test:e2e:run": "export NODE_OPTIONS=--openssl-legacy-provider && wait-on http://localhost:5001/reservations?date=2020-01-01 && jest --detectOpenHandles --forceExit --config ./e2e/jest.config.js",
+        "test:start:backend": "npm run test:start --prefix ./../back-end",
+        "test:start:frontend": "npx cross-env BROWSER=none NODE_ENV=test react-scripts --openssl-legacy-provider start",
+        "test:unit": "react-scripts test"
+    },
+    "eslintConfig": {
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "devDependencies": {
+        "cross-fetch": "^3.0.6",
+        "jest-puppeteer": "^6.0.0",
+        "npm-run-all": "^4.1.5",
+        "prettier": "3.2.4",
+        "puppeteer": "^10.4.0"
+    }
 }
