@@ -8,7 +8,7 @@ export const FormSelect = ({ labelHtml, label, data, ...rest }) => {
             <select {...rest}>
                 <option value=''>-- Select a Table --</option>
 
-                {data.map((table) => (
+                {data.data.map((table) => (
                     <option key={table.table_id} value={table.table_id}>
                         {table.table_name} - {table.capacity}
                     </option>
